pyproject.toml
# Copyright (C) 2026 Julian Valentin
#
# This Source Code Form is subject to the terms of the Mozilla Public
# License, v. 2.0. If a copy of the MPL was not distributed with this
# file, You can obtain one at https://mozilla.org/MPL/2.0/.

[build-system]
requires = ["uv_build>=0.11.0,<0.12.0"]
build-backend = "uv_build"

[project]
name = "stupid-backup"
version = "1.0.0"
description = "Stupid backup tool"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "ruff>=0.15.9",
    "ty>=0.0.29",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
stupid-backup = "stupid_backup:main.main"

[tool.mypy]
files = ["src"]
strict = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = [
    "COM812",
    "D203",
    "D213",
    "S310",
    "S311",
    "S603",
    "S607",
]
select = ["ALL"]
